+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - FIX query planner did not pick the correct index on `$eq` operations.
 - IMPROVE performance of the memory RxStorage
+- IMPROVE performance of custom index creation
 
 <!-- ADD new changes here! -->